// src/app/features/shared/models/notification/notification-type.enum.ts
export enum NotificationType {
  ACHIEVEMENT_UNLOCKED = 'ACHIEVEMENT_UNLOCKED',
  ITEM_ACQUIRED = 'ITEM_ACQUIRED',
  LEVEL_UP = 'LEVEL_UP',
  GROUP_INVITATION = 'GROUP_INVITATION',
  GROUP_ITEM_USED = 'GROUP_ITEM_USED',
  NEW_GROUP_MEMBER = 'NEW_GROUP_MEMBER',
  GROUP_MEMBER_LEFT = 'GROUP_MEMBER_LEFT',
  NEW_GROUP_MESSAGE = 'NEW_GROUP_MESSAGE',
  GROUP_TASK_ASSIGNED = 'GROUP_TASK_ASSIGNED',
  GROUP_TASK_COMPLETED = 'GROUP_TASK_COMPLETED',
  GROUP_REQUEST_STATUS_UPDATED = 'GROUP_REQUEST_STATUS_UPDATED',
  NEW_GROUP_REQUEST = 'NEW_GROUP_REQUEST',
  OTHER = 'OTHER',
}